PE-7199 Refactor config and imports
