--- .babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env", {
-        "useBuiltIns": "entry"
-      }
-    ]
-  ]
-}
