+++ src/config/config.js
@@ -7,6 +7,16 @@ const config = {
     release_version: process.env.MOCK_API_RELEASE_VERSION || '',
     start_time: Date.now(),
   },
+  drakovOptions: {
+    sourceFiles: './blueprints/*.apib',
+    serverPort: parseInt(process.env.MOCK_API_PORT || 3000, 10),
+    method: ['GET', 'PATCH', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+    debugMode: false,
+    stealthmode: false,
+    disableCORS: false,
+    discover: false,
+    watch: false,
+  },
 };
 
 module.exports = config;
