+++ src/routes/index.js
@@ -7,9 +7,7 @@ const routes = {};
 
 fs
   .readdirSync(__dirname)
-  .filter((file) => {
-    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
-  })
+  .filter(file => (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js'))
   .forEach((file) => {
     const fullPath = path.join(__dirname, file);
     const fileName = file.substring(0, file.length - 3);
